refactor: payment 관련 uri 수정
#139
