+++ api/src/main/kotlin/com/inner/circle/api/controller/PaymentClaimController.kt
@@ -1,16 +1,28 @@
 package com.inner.circle.api.controller
 
-import com.inner.circle.api.structure.dto.PaymentResponse
+import com.inner.circle.api.controller.dto.ConfirmPaymentDto
+import com.inner.circle.api.controller.dto.PaymentApproveDto
+import com.inner.circle.api.controller.dto.PaymentResponse
+import com.inner.circle.api.controller.request.ConfirmPaymentRequest
+import com.inner.circle.api.controller.request.ConfirmSimplePaymentRequest
+import com.inner.circle.api.controller.request.PaymentApproveRequest
+import com.inner.circle.core.usecase.ConfirmPaymentUseCase
+import com.inner.circle.core.usecase.ConfirmSimplePaymentUseCase
 import com.inner.circle.core.usecase.PaymentClaimUseCase
+import com.inner.circle.core.usecase.SavePaymentApproveUseCase
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 
 @Tag(name = "Claims", description = "Claim (결제 요청) API")
 @PaymentV1Api
 class PaymentClaimController(
-    private val claimUseCase: PaymentClaimUseCase
+    private val confirmPaymentUseCase: ConfirmPaymentUseCase,
+    private val claimUseCase: PaymentClaimUseCase,
+    private val savePaymentApproveService: SavePaymentApproveUseCase
 ) {
     @Operation(summary = "결제 요청")
     @PostMapping("/payments")
@@ -20,5 +32,79 @@ class PaymentClaimController(
         val merchantId = "tempMerchantId" // @AuthenticationPrincipal
         val response = claimUseCase.createPayment(request, merchantId)
         return PaymentResponse.ok(response)
+    }
+
+    @Operation(summary = "결제 승인 API")
+    @PostMapping("/payments/confirm")
+    fun confirmPayment(
+        @RequestBody paymentApproveRequest: PaymentApproveRequest
+    ): PaymentResponse<PaymentApproveDto> =
+        PaymentResponse.ok(
+            savePaymentApproveService
+                .saveApprove(
+                    SavePaymentApproveUseCase.Request(
+                        paymentKey = paymentApproveRequest.paymentKey,
+                        orderId = paymentApproveRequest.orderId,
+                        amount = paymentApproveRequest.amount
+                    )
+                ).let {
+                    PaymentApproveDto(
+                        orderId = it.orderId,
+                        paymentKey = it.paymentKey,
+                        amount = it.amount
+                    )
+                }
+        )
+
+    @Operation(summary = "간편 결제 확인")
+    @PostMapping("/payments/orders/{order_id}/simple")
+    fun proceedPaymentConfirm(
+        @PathVariable("order_id") orderId: String,
+        @RequestBody confirmSimplePaymentRequest: ConfirmSimplePaymentRequest
+    ): PaymentResponse<ConfirmPaymentDto> {
+        val response =
+            PaymentResponse.ok(
+                ConfirmPaymentDto.of(
+                    confirmPaymentUseCase.confirmPayment(
+                        ConfirmSimplePaymentUseCase.Request(
+                            orderId = orderId,
+                            merchantId = confirmSimplePaymentRequest.merchantId
+                        )
+                    )
+                )
+            )
+        return response
+    }
+
+    @Operation(summary = "결제 확인")
+    @PostMapping("/payments/orders/{order_id}")
+    fun proceedPaymentConfirm(
+        @PathVariable("order_id") orderId: String,
+        @RequestBody confirmPaymentRequest: ConfirmPaymentRequest
+    ): PaymentResponse<ConfirmPaymentDto> {
+        val response =
+            PaymentResponse.ok(
+                ConfirmPaymentDto.of(
+                    confirmPaymentUseCase.confirmPayment(
+                        ConfirmPaymentUseCase.Request(
+                            orderId = orderId,
+                            merchantId = confirmPaymentRequest.merchantId,
+                            cardNumber = confirmPaymentRequest.cardNumber,
+                            expirationPeriod = confirmPaymentRequest.expirationPeriod,
+                            cvc = confirmPaymentRequest.cvc
+                        )
+                    )
+                )
+            )
+        return response
+    }
+
+    @Operation(summary = "결제 취소")
+    @GetMapping("/payments/orders/{order_id}/cancel")
+    fun cancelPaymentConfirm(
+        @PathVariable("order_id") orderId: String
+    ): PaymentResponse<String> {
+        val response = PaymentResponse.ok("결제가 취소되었습니다.")
+        return response
     }
 }
