+++ api/src/main/kotlin/com/inner/circle/api/controller/request/ConfirmPaymentRequest.kt
@@ -1,7 +1,6 @@
 package com.inner.circle.api.controller.request
 
 data class ConfirmPaymentRequest(
-    val orderId: String,
     val merchantId: String,
     val cardNumber: String,
     val expirationPeriod: String,
