+++ api/src/main/kotlin/com/inner/circle/api/controller/request/ConfirmSimplePaymentRequest.kt
@@ -1,6 +1,5 @@
 package com.inner.circle.api.controller.request
 
 data class ConfirmSimplePaymentRequest(
-    val orderId: String,
     val merchantId: String
 )
